Improve BitArray512 tests, remove useless cast

+++ src/OpenSage.Mathematics/BitArray512.cs
@@ -84,7 +84,7 @@ namespace OpenSage.Mathematics
                 }
                 else
                 {
-                    *(pointer + offset) &= (byte) ~mask;
+                    *(pointer + offset) &= ~mask;
                 } 
             }
